Fix compiler nag

+++ source/MrsWatson.c
@@ -351,7 +351,7 @@ int mrsWatsonMain(ErrorReporter errorReporter, int argc, char** argv) {
           setBlocksize((const unsigned long)programOptionsGetNumber(programOptions, OPTION_BLOCKSIZE));
           break;
         case OPTION_CHANNELS:
-          setNumChannels((const unsigned long)programOptionsGetNumber(programOptions, OPTION_CHANNELS));
+          setNumChannels((const unsigned int)programOptionsGetNumber(programOptions, OPTION_CHANNELS));
           break;
         case OPTION_DISPLAY_INFO:
           shouldDisplayPluginInfo = true;
